Restructure trustnet files
