+++ pkg/trustnet/detector_test.go
